Video 46

+++ BlazorTutorial/EmployeeManagement.Web/Pages/DisplayEmployeeBase.cs
@@ -1,10 +1,10 @@
-﻿using Microsoft.AspNetCore.Components;
+﻿using EmployeeManagement.Models;
+using EmployeeManagement.Web.Services;
+using Microsoft.AspNetCore.Components;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using EmployeeManagement.Models;
-
 
 namespace EmployeeManagement.Web.Pages
 {
@@ -15,12 +15,10 @@ namespace EmployeeManagement.Web.Pages
 
         [Parameter]
         public bool ShowFooter { get; set; }
-        protected bool IsSelected { get; set; }
 
         [Parameter]
         public EventCallback<bool> OnEmployeeSelection { get; set; }
 
-        
         [Parameter]
         public EventCallback<int> OnEmployeeDeleted { get; set; }
 
@@ -30,15 +28,24 @@ namespace EmployeeManagement.Web.Pages
         [Inject]
         public NavigationManager NavigationManager { get; set; }
 
-        protected async Task Delete_Click()
+        protected PragimTech.Components.ConfirmBase DeleteConfirmation { get; set; }
+
+        protected void Delete_Click()
         {
-            await EmployeeService.DeleteEmployee(Employee.EmployeeId);
-            await OnEmployeeDeleted.InvokeAsync(Employee.EmployeeId);
-            //NavigationManager.NavigateTo("/", true);
+            DeleteConfirmation.Show();
         }
+
+        protected async Task ConfirmDelete_Click(bool deleteConfirmed)
+        {
+            if (deleteConfirmed)
+            {
+                await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+                await OnEmployeeDeleted.InvokeAsync(Employee.EmployeeId);
+            }
+        }
+
         protected async Task CheckBoxChanged(ChangeEventArgs e)
         {
-            IsSelected = (bool)e.Value;
             await OnEmployeeSelection.InvokeAsync((bool)e.Value);
         }
     }
